- fix some warnings on shadow's driver
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@7675 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 4a43573d6b454d3310d118a4a859850f26879314
Former-commit-id: 7ea2853fcdf017929d7f99bc3d4fd625a28da0b3

+++ src/drivers/shadow/src/SpringsPath.cpp
@@ -722,7 +722,7 @@ void	SpringsPath::OptimisePtClothoid(
 
 void SpringsPath::ResetSpringVelocities()
 {
-    if( m_temp.size() != NSEG )
+    if( m_temp.size() != (unsigned int)NSEG )
     {
         m_temp.resize( NSEG );
     }
@@ -749,7 +749,7 @@ void SpringsPath::OptimisePathSprings(
 {
     const int	NSEG = m_pTrack->GetSize();
 
-    if( m_temp.size() != NSEG )
+    if( m_temp.size() != (unsigned int)NSEG )
     {
         m_temp.resize( NSEG );
         ResetSpringVelocities();
